refactor: extract LegendTab from ChartPropertiesPanel
